Add support for drop-in files
The drop-in directory must not created more then once, even if multiple
drop-in files per unit are created. Also drop-in files for units, which are
not deployed via puppet, should be supported. The only solution is to use
ensure_resource from puppetlabs/stdlib. Everything else is the same as for a
unit_file, except the name of the drop-in file, which has to end with '.conf'.