Implement a workaround for PUP-9473.
Prior to this commit, the follow code did not suffer from PUP-9473:

    systemd::unit_file { 'myservice.service':
      ensure => absent,
      active => true,
    }

That's because the module considers that an invalid state and fails to
compile.

The follow code did trigger the bug:

    systemd::unit_file { 'myservice.service':
      ensure => absent,
    }

    service { 'myservice':
      ensure  => running,
      require => Systemd::Unit_file['myservice.service'],
    }

That's precisely what happens when a module switches from a
module-provided unit file to a package-provided unit file.