systemd::escape: Also escape - (dash)
We also need to escape dashes in unit names, for instance for .swap
services on device nodes with dashes in them.

systemd-escape escapes dashes:
  # systemd-escape -p /dev/disk/by-id/scsi-0QEMU_QEMU_HARDDISK_drive-ua-swap
  dev-disk-by\x2did-scsi\x2d0QEMU_QEMU_HARDDISK_drive\x2dua\x2dswap
but this function does not (yet).

In the long term, we may want to more closely match the behaviour of
systemd's unit-name.c, but this is all I need for now.

(Also cf. #242.)