Support Service Limits specified in Bytes
Previously the ServiteLimits type restricted limits specified in bytes to contain
a multiplicative suffix. The systemd man pages ([0], [1]) specify these are optional
and a bare integer can be specified, which will be interpreted as Bytes.

This commit makes the multiplicative suffixes optional, and adds tests for two cases
to ensure non-suffixed values are accepted.

[0] https://www.freedesktop.org/software/systemd/man/systemd.exec.html
    (For `Limit*` options)
[1] https://www.freedesktop.org/software/systemd/man/systemd.resource-control.html#
    (For `Memory*` options)