Add restart_service parameter on service_limits for compatibility
In 8a9e03a4d6a58a5a5083d7ead60dcbc87f727d2f the parameter was removed,
but this makes it very hard for modules to be compatible with both < 4
and >= 4. Requiring >= 4 will be painful, so this adds the parameter
again but only accepts the parameter false.

A concrete example is the puppet-redis module which sets restart_service
to false to prevent a restart. This would be compatible with both the
old and new behavior, making upgrades smoother.

Fixes: 8a9e03a4d6a58a5a5083d7ead60dcbc87f727d2f