Addition of timer directives to manage_unit and manage_dropin
The timer directives for unit files is now complete for options
present in the man page of systemd-253.

For example:

```puppet
systemd::manage_unit{'my.timer':
  ensure => present,
  unit_entry => {
    'Description' => 'my timer',
  },
  timer_entry => {
    'Calendar' => 'daily',
    'AccuracySecs' => '1 hour',
  },
}
```

Previously only the 'Calendar' directive was available and this
patch also includes a bug fix for that.

Previously

```
systemd::manage_dropin{'my.timer':
  ensure => present,
  unit => 'myunit.service',
  timer_entry => {
    'Calendar' => ['','daily'],
  },
}
```
with an empty resetting first `''` element was not permitted and now it is..