feat: add a cron like wrapper for timers
This helper will allow to define a timer in a cron like way, and will
setup a corresponding timer and service.
In basic usage it only requires a command, user, timerspec and an ensure
parameter.
For additional features the parameter `service_overrides`, `timer_overrides`,
`service_unit_overrides` and `timer_unit_overrides` can be used to define
additional aspects of the create units.
The correct order of the units is also ensured.

fixes #374
related to https://github.com/voxpupuli/puppet-systemd/pull/287