Use runuser or run0 to connect to systemd --user instance
Create a new function `systemd::systemctl_user` to construct
array of command used to contact `systemd --user` instance so
`systemctl --user status my.service` can be executed.

In addition we switch to using `runuser` or `run0` for systemd >= 256 to become the user.

* Using `systemctl` or `systemd-run` with --machinectl $user@.host
  always fails when puppet is ran inside its own systemd unit.
  I believe systemd services can not access `systemd --user` instances.
* We use `runuser -u $user` rather than a `user => $user` on the exec
  since the $(id -u $user) must be resolved late within the exec itself.

Now the command is generated inside a function this will make future
improvements to this method much simpler.